perf: 修改 探测
